fixed index buffers

+++ sadx-d3d11/d3d8to9_index_buffer.cpp
@@ -158,7 +158,7 @@ HRESULT STDMETHODCALLTYPE Direct3DIndexBuffer8::Lock(UINT OffsetToLock, UINT Siz
 		return D3DERR_INVALIDCALL;
 	}
 
-	if (OffsetToLock + SizeToLock > buffer.size())
+	if ((uint64_t)OffsetToLock + (uint64_t)SizeToLock > (uint64_t)buffer.size())
 	{
 		return D3DERR_INVALIDCALL;
 	}
